🔨➖[Devlopment,Dependency] Edits SaveNoteServelet to capture form data
So we downgraded the tomcat and teh SaveNoteServlet error was fixed. Tomcat 10 was a huge pain.The whole things was not working and throwing classnotfound exception.

Now its fixed and we removed the servelet dependency since it is not needed really.
We edited SaveNoteServelet and add_notes.jsp (added name tag to form field) to capture data from the form.